# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Facebook AI Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List all AI providers
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
    post:
      operationId: createProvider
      tags: [providers]
      summary: Create a new AI provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderInput"
      responses:
        "201":
          description: Created provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
  /providers/{id}:
    put:
      operationId: updateProvider
      tags: [providers]
      summary: Update an AI provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProviderInput"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
    delete:
      operationId: deleteProvider
      tags: [providers]
      summary: Delete an AI provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /providers/{id}/activate:
    post:
      operationId: activateProvider
      tags: [providers]
      summary: Set a provider as the active one
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /providers/{id}/test:
    post:
      operationId: testProvider
      tags: [providers]
      summary: Test an AI provider connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderTestResult"
...
  /slots:
    get:
      operationId: listSlots
      tags: [slots]
      summary: List all time slots
      responses:
        "200":
          description: List of slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailableSlot"
    post:
      operationId: createSlot
      tags: [slots]
      summary: Create a new time slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSlotInput"
      responses:
        "201":
          description: Created slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlot"
  /slots/{id}:
    patch:
      operationId: updateSlot
      tags: [slots]
      summary: Update a slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSlotInput"
      responses:
        "200":
          description: Updated slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlot"
    delete:
      operationId: deleteSlot
      tags: [slots]
      summary: Delete a slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /slots/available:
    get:
      operationId: getAvailableSlots
      tags: [slots]
      summary: Get available slots for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available slots with booking capacity info
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/AvailableSlot"
                    - type: object
                      properties:
                        currentBookings:
                          type: integer
                        remainingCapacity:
                          type: integer

  # ── FAQs ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ────────────────────────────────────────────────────────────────
...
    Provider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        providerType:
          type: string
        apiKey:
          type: string
          description: Masked key (never returned in full)
        baseUrl:
          type: string
          nullable: true
        modelName:
          type: string
        isActive:
          type: integer
        priority:
          type: integer
          description: Round-robin priority (lower = higher priority)
        isEnabled:
          type: integer
          description: Whether provider participates in load balancing (0 or 1)
        failCount:
          type: integer
          description: Consecutive failure count (resets on success)
        lastUsedAt:
          type: string
          nullable: true
          description: ISO timestamp of last use
        createdAt:
          type: string
      required: [id, name, providerType, modelName, isActive]
    CreateProviderInput:
      type: object
      properties:
        name:
          type: string
        providerType:
          type: string
        apiKey:
          type: string
        baseUrl:
          type: string
          nullable: true
        modelName:
          type: string
      required: [name, providerType, apiKey, modelName]
    UpdateProviderInput:
      type: object
      properties:
        name:
          type: string
        providerType:
          type: string
        apiKey:
          type: string
        baseUrl:
          type: string
          nullable: true
        modelName:
          type: string
        priority:
          type: integer
        isEnabled:
          type: integer
    ProviderTestResult:
      type: object
      properties:
        success:
          type: boolean
        response:
          type: string
        latencyMs:
          type: integer
      required: [success]
...
    AvailableSlot:
      type: object
      properties:
        id:
          type: integer
        dayOfWeek:
          type: integer
        timeSlot:
          type: string
        isActive:
          type: integer
        maxBookings:
          type: integer
      required: [id, dayOfWeek, timeSlot, isActive, maxBookings]
    CreateSlotInput:
      type: object
      properties:
        dayOfWeek:
          type: integer
        timeSlot:
          type: string
        maxBookings:
          type: integer
      required: [dayOfWeek, timeSlot]
    UpdateSlotInput:
      type: object
      properties:
        isActive:
          type: integer
        maxBookings:
          type: integer

    # ── FAQs ──────────────────────────────────────────────────────────────────
...